* master: (195 commits)
  [SPARK-8739] [WEB UI] [WINDOWS] A illegal character `\r` can be contained in StagePage.
  [SPARK-8563] [MLLIB] Fixed a bug so that IndexedRowMatrix.computeSVD().U.numCols = k
  [SPARK-8705] [WEBUI] Don't display rects when totalExecutionTime is 0
  [SPARK-8736] [ML] GBTRegressor should not threshold prediction
  [SPARK-8372] Do not show applications that haven't recorded their app ID yet.
  [SPARK-2645] [CORE] Allow SparkEnv.stop() to be called multiple times without side effects.
  [SPARK-8560] [UI] The Executors page will have negative if having resubmitted tasks
  [SPARK-7514] [MLLIB] Add MinMaxScaler to feature transformation
  [SPARK-8471] [ML] Discrete Cosine Transform Feature Transformer
  [SPARK-8628] [SQL] Race condition in AbstractSparkSQLParser.parse
  [SPARK-8664] [ML] Add PCA transformer
  [SPARK-6785] [SQL] fix DateTimeUtils for dates before 1970
  [SPARK-8619] [STREAMING] Don't recover keytab and principal configuration within Streaming checkpoint
  [SPARK-8630] [STREAMING] Prevent from checkpointing QueueInputDStream
  [SPARK-7988] [STREAMING] Round-robin scheduling of receivers by default
  [SPARK-8615] [DOCUMENTATION] Fixed Sample deprecated code
  [SPARK-8713] Make codegen thread safe
  [SPARK-8679] [PYSPARK] [MLLIB] Default values in Pipeline API should be immutable
  [SPARK-4127] [MLLIB] [PYSPARK] Python bindings for StreamingLinearRegressionWithSGD
  [SPARK-8437] [DOCS] Corrected: Using directory path without wildcard for filename slow for large number of files with wholeTextFiles and binaryFiles
  ...

+++ core/src/main/scala/org/apache/spark/SparkEnv.scala
@@ -22,7 +22,6 @@ import java.net.Socket
 
 import akka.actor.ActorSystem
 
-import scala.collection.JavaConversions._
 import scala.collection.mutable
 import scala.util.Properties
 
@@ -90,39 +89,42 @@ class SparkEnv (
   private var driverTmpDirToDelete: Option[String] = None
 
   private[spark] def stop() {
-    isStopped = true
-    pythonWorkers.foreach { case(key, worker) => worker.stop() }
-    Option(httpFileServer).foreach(_.stop())
-    mapOutputTracker.stop()
-    shuffleManager.stop()
-    broadcastManager.stop()
-    blockManager.stop()
-    blockManager.master.stop()
-    metricsSystem.stop()
-    outputCommitCoordinator.stop()
-    rpcEnv.shutdown()
 
-    // Unfortunately Akka's awaitTermination doesn't actually wait for the Netty server to shut
-    // down, but let's call it anyway in case it gets fixed in a later release
-    // UPDATE: In Akka 2.1.x, this hangs if there are remote actors, so we can't call it.
-    // actorSystem.awaitTermination()
+    if (!isStopped) {
+      isStopped = true
+      pythonWorkers.values.foreach(_.stop())
+      Option(httpFileServer).foreach(_.stop())
+      mapOutputTracker.stop()
+      shuffleManager.stop()
+      broadcastManager.stop()
+      blockManager.stop()
+      blockManager.master.stop()
+      metricsSystem.stop()
+      outputCommitCoordinator.stop()
+      rpcEnv.shutdown()
 
-    // Note that blockTransferService is stopped by BlockManager since it is started by it.
+      // Unfortunately Akka's awaitTermination doesn't actually wait for the Netty server to shut
+      // down, but let's call it anyway in case it gets fixed in a later release
+      // UPDATE: In Akka 2.1.x, this hangs if there are remote actors, so we can't call it.
+      // actorSystem.awaitTermination()
 
-    // If we only stop sc, but the driver process still run as a services then we need to delete
-    // the tmp dir, if not, it will create too many tmp dirs.
-    // We only need to delete the tmp dir create by driver, because sparkFilesDir is point to the
-    // current working dir in executor which we do not need to delete.
-    driverTmpDirToDelete match {
-      case Some(path) => {
-        try {
-          Utils.deleteRecursively(new File(path))
-        } catch {
-          case e: Exception =>
-            logWarning(s"Exception while deleting Spark temp dir: $path", e)
+      // Note that blockTransferService is stopped by BlockManager since it is started by it.
+
+      // If we only stop sc, but the driver process still run as a services then we need to delete
+      // the tmp dir, if not, it will create too many tmp dirs.
+      // We only need to delete the tmp dir create by driver, because sparkFilesDir is point to the
+      // current working dir in executor which we do not need to delete.
+      driverTmpDirToDelete match {
+        case Some(path) => {
+          try {
+            Utils.deleteRecursively(new File(path))
+          } catch {
+            case e: Exception =>
+              logWarning(s"Exception while deleting Spark temp dir: $path", e)
+          }
         }
+        case None => // We just need to delete tmp dir created by driver, so do nothing on executor
       }
-      case None => // We just need to delete tmp dir created by driver, so do nothing on executor
     }
   }
